Refactor: 테스트 구조 변경
