+++ build.gradle
@@ -70,6 +70,7 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 
 	// test
+	implementation 'javax.xml.bind:jaxb-api:2.3.1'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.assertj:assertj-core:3.23.1'
 	testImplementation 'org.springframework.security:spring-security-test'
