+++ src/test/java/com/cojac/storyteller/unit/page/PageServiceUnitTest.java
@@ -1,0 +1,143 @@
+package com.cojac.storyteller.unit.page;
+
+import com.cojac.storyteller.book.entity.BookEntity;
+import com.cojac.storyteller.book.exception.BookNotFoundException;
+import com.cojac.storyteller.book.repository.BookRepository;
+import com.cojac.storyteller.page.dto.PageDetailResponseDTO;
+import com.cojac.storyteller.page.entity.PageEntity;
+import com.cojac.storyteller.page.exception.PageNotFoundException;
+import com.cojac.storyteller.page.repository.PageRepository;
+import com.cojac.storyteller.page.service.PageService;
+import com.cojac.storyteller.profile.entity.ProfileEntity;
+import com.cojac.storyteller.profile.exception.ProfileNotFoundException;
+import com.cojac.storyteller.profile.repository.ProfileRepository;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.*;
+
+/**
+ * 단위 테스트 클래스
+ *
+ * 이 클래스는 개별 단위(주로 서비스 또는 비즈니스 로직) 기능을 검증하기 위한 단위 테스트를 포함합니다.
+ *
+ * 주요 특징:
+ * - 모의 객체(mock objects)를 사용하여 외부 의존성을 제거하고,테스트 대상 객체의 로직에만 집중합니다.
+ * - 테스트의 독립성을 보장하여, 각 테스트가 서로에게 영향을 미치지 않도록 합니다.
+ *
+ * 테스트 전략:
+ * - 간단한 기능이나 로직에 대한 테스트는 단위 테스트를 사용하십시오.
+ * - 시스템의 전체적인 동작 및 상호작용을 검증하기 위해 통합 테스트를 활용하십시오.
+ *
+ * 참고: 단위 테스트는 실행 속도가 빠르며,
+ *       전체 시스템의 동작보다는 개별 단위의 동작을 검증하는 데 중점을 둡니다.
+ */
+@ExtendWith(MockitoExtension.class)
+class PageServiceUnitTest {
+
+    @Mock
+    private PageRepository pageRepository;
+
+    @Mock
+    private BookRepository bookRepository;
+
+    @Mock
+    private ProfileRepository profileRepository;
+
+    @InjectMocks
+    private PageService pageService;
+
+    private ProfileEntity profile;
+    private BookEntity book;
+    private PageEntity page;
+
+    @BeforeEach
+    void setUp() {
+        profile = ProfileEntity.builder().id(1).build();
+        book = BookEntity.builder().id(1).profile(profile).title("Test Book").build();
+        page = PageEntity.builder().id(1).pageNumber(1).image("imageUrl").content("Page content").build();
+    }
+
+    /**
+     * 페이지 세부 정보 가져오기
+     */
+    @Test
+    @DisplayName("페이지 세부 정보 가져오기 단위 테스트 - 성공")
+    void testGetPageDetail_Success() {
+        // given
+        Integer profileId = 1;
+        Integer bookId = 1;
+        Integer pageNum = 1;
+
+        when(profileRepository.findById(profileId)).thenReturn(Optional.of(profile));
+        when(bookRepository.findByIdAndProfile(bookId, profile)).thenReturn(Optional.of(book));
+        when(pageRepository.findPageWithUnknownWords(book, pageNum)).thenReturn(Optional.of(page));
+
+        // when
+        PageDetailResponseDTO result = pageService.getPageDetail(profileId, bookId, pageNum);
+
+        // then
+        assertNotNull(result);
+        assertEquals(page.getId(), result.getPageId());
+        assertEquals(page.getPageNumber(), result.getPageNumber());
+        assertEquals(page.getImage(), result.getImage());
+        assertEquals(page.getContent(), result.getContent());
+
+        verify(profileRepository, times(1)).findById(profileId);
+        verify(bookRepository, times(1)).findByIdAndProfile(bookId, profile);
+        verify(pageRepository, times(1)).findPageWithUnknownWords(book, pageNum);
+    }
+
+    @Test
+    @DisplayName("페이지 세부 정보 가져오기 단위 테스트 - 프로필이 존재하지 않을 때 예외 처리")
+    void testGetPageDetail_ProfileNotFound() {
+        // given
+        Integer profileId = 1;
+        Integer bookId = 1;
+        Integer pageNum = 1;
+
+        when(profileRepository.findById(profileId)).thenReturn(Optional.empty());
+
+        // when & then
+        assertThrows(ProfileNotFoundException.class, () -> pageService.getPageDetail(profileId, bookId, pageNum));
+    }
+
+    @Test
+    @DisplayName("페이지 세부 정보 가져오기 단위 테스트 - 책이 존재하지 않을 때 예외 처리")
+    void testGetPageDetail_BookNotFound() {
+        // given
+        Integer profileId = 1;
+        Integer bookId = 1;
+        Integer pageNum = 1;
+
+        when(profileRepository.findById(profileId)).thenReturn(Optional.of(profile));
+        when(bookRepository.findByIdAndProfile(bookId, profile)).thenReturn(Optional.empty());
+
+        // when & then
+        assertThrows(BookNotFoundException.class, () -> pageService.getPageDetail(profileId, bookId, pageNum));
+    }
+
+    @Test
+    @DisplayName("페이지 세부 정보 가져오기 단위 테스트 - 페이지가 존재하지 않을 때 예외 처리")
+    void testGetPageDetail_PageNotFound() {
+        // given
+        Integer profileId = 1;
+        Integer bookId = 1;
+        Integer pageNum = 1;
+
+        when(profileRepository.findById(profileId)).thenReturn(Optional.of(profile));
+        when(bookRepository.findByIdAndProfile(bookId, profile)).thenReturn(Optional.of(book));
+        when(pageRepository.findPageWithUnknownWords(book, pageNum)).thenReturn(Optional.empty());
+
+        // when & then
+        assertThrows(PageNotFoundException.class, () -> pageService.getPageDetail(profileId, bookId, pageNum));
+    }
+}
