+++ src/test/java/com/cojac/storyteller/unit/user/security/CustomUserDetailsServiceTest.java
@@ -1,0 +1,66 @@
+package com.cojac.storyteller.unit.user.security;
+
+import com.cojac.storyteller.user.entity.LocalUserEntity;
+import com.cojac.storyteller.user.repository.LocalUserRepository;
+import com.cojac.storyteller.user.service.security.CustomUserDetailsService;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.when;
+
+/**
+ * 단위 테스트 클래스
+ *
+ * 이 클래스는 개별 단위(주로 서비스 또는 비즈니스 로직) 기능을 검증하기 위한 단위 테스트를 포함합니다.
+ *
+ * 주요 특징:
+ * - 모의 객체(mock objects)를 사용하여 외부 의존성을 제거하고, 테스트 대상 객체의 로직에만 집중합니다.
+ * - 테스트의 독립성을 보장하여, 각 테스트가 서로에게 영향을 미치지 않도록 합니다.
+ *
+ * 테스트 전략:
+ * - 간단한 기능이나 로직에 대한 테스트는 단위 테스트를 사용하십시오.
+ * - 시스템의 전체적인 동작 및 상호작용을 검증하기 위해 통합 테스트를 활용하십시오.
+ *
+ * 참고: 단위 테스트는 실행 속도가 빠르며,
+ *       전체 시스템의 동작보다는 개별 단위의 동작을 검증하는 데 중점을 둡니다.
+ */
+@ExtendWith(MockitoExtension.class)
+public class CustomUserDetailsServiceTest {
+
+    @InjectMocks
+    private CustomUserDetailsService customUserDetailsService;
+
+    @Mock
+    private LocalUserRepository localUserRepository;
+
+    @Test
+    @DisplayName("사용자가 존재할 때 사용자 세부정보를 로드해야 합니다.")
+    public void testLoadUserByUsername_UserExists() {
+        LocalUserEntity mockUser = new LocalUserEntity("user", "password", "email", "ROLE_USER");
+        when(localUserRepository.findByUsername("user")).thenReturn(Optional.of(mockUser));
+
+        UserDetails userDetails = customUserDetailsService.loadUserByUsername("user");
+
+        assertNotNull(userDetails);
+        assertEquals("user", userDetails.getUsername());
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 사용자에 대해 UsernameNotFoundException을 발생시켜야 합니다.")
+    public void testLoadUserByUsername_UserNotFound() {
+        when(localUserRepository.findByUsername("unknown")).thenReturn(Optional.empty());
+
+        assertThrows(UsernameNotFoundException.class, () -> {
+            customUserDetailsService.loadUserByUsername("unknown");
+        });
+    }
+}
